Move raytrace func to plugin, coerce non rgb to rgb, tweak optimizations
