+++ coloraide/algebra.py
@@ -677,117 +677,6 @@ def interpolate(points: list[Vector], method: str = 'linear') -> Interpolate:
 
 
 ################################
-# Ray tracing.
-################################
-def raytrace_cube(size: Vector, start: Vector, end: Vector) -> tuple[int, Vector]:
-    """
-    Returns the face and the intersection tuple of an array from start to end of a cube of size [x, y, z].
-
-    - 0: None, (point is None)
-    - 1: intersection with x==0 face,
-    - 2: intersection with x==size[0] face,
-    - 3: intersection with y==0 face,
-    - 4: intersection with y==size[1] face,
-    - 5: intersection with z==0 face,
-    - 6: intersection with z==size[2] face,
-
-    The cube is an axis-aligned cube: (0,0,0)-(size[0],size[1],size[2]).
-
-    ```
-    mwt (https://math.stackexchange.com/users/591865/mwt),
-    Finding the side of a cube intersecting a line using the shortest computation,
-    URL (version: 2020-08-01): https://math.stackexchange.com/q/3776157
-    ```
-    """
-
-    # Negated deltas
-    ndx = start[0] - end[0]
-    ndy = start[1] - end[1]
-    ndz = start[2] - end[2]
-
-    # Sizes scaled by the negated deltas
-    sxy = ndx * size[1]
-    sxz = ndx * size[2]
-    syx = ndy * size[0]
-    syz = ndy * size[2]
-    szx = ndz * size[0]
-    szy = ndz * size[1]
-
-    # Cross terms
-    cxy = end[0] *start[1] - end[1] * start[0]
-    cxz = end[0] *start[2] - end[2] * start[0]
-    cyz = end[1] *start[2] - end[2] * start[1]
-
-    # Absolute delta products
-    axy = abs(ndx * ndy)
-    axz = abs(ndx * ndz)
-    ayz = abs(ndy * ndz)
-    axyz = abs(ndz * axy)
-
-    # Default to "no intersection"
-    face_num = 0
-    face_tau = abs(ndz * axy)
-
-    # These variables are no longer used:
-    del ndx, ndy, ndz
-
-    if start[0] < 0 and 0 < end[0]:
-        # Face 1: x == 0
-        tau = -start[0] * ayz
-        if tau < face_tau and cxy >= 0 and cxz >= 0 and cxy <= -sxy and cxz <= -sxz:
-            face_tau = tau
-            face_num = 1
-
-    elif end[0] < size[0] and size[0] < start[0]:
-        # Face 2: x == size[0]
-        tau = (start[0] - size[0]) * ayz
-        if tau < face_tau and cxy <= syx and cxz <= szx and cxy >= syx - sxy and cxz >= szx - sxz:
-            face_tau = tau
-            face_num = 2
-
-    if start[1] < 0 and end[1] > 0:
-        # Face 3: y == 0
-        tau = -start[1] * axz
-        if tau < face_tau and cxy <= 0 and cyz >= 0 and cxy >= syx and cyz <= -syz:
-            face_tau = tau
-            face_num = 3
-
-    elif start[1] > size[1] and end[1] < size[1]:
-        # Face 4: y == size[1]
-        tau = (start[1] - size[1]) * axz
-        if tau < face_tau and cxy >= -sxy and cyz <= szy and cxy <= syx - sxy and cyz >= szy - syz:
-            face_tau = tau
-            face_num = 4
-
-    if start[2] < 0 and end[2] > 0:
-        # Face 5: z == 0
-        tau = -start[2] * axy
-        if tau < face_tau and cxz <= 0 and cyz <= 0 and cxz >= szx and cyz >= szy:
-            face_tau = tau
-            face_num = 5
-
-    elif start[2] > size[2] and end[2] < size[2]:
-        # Face 6: z == size[2]
-        tau = (start[2] - size[2]) * axy
-        if tau < face_tau and cxz >= -sxz and cyz >= -syz and cxz <= szx - sxz and cyz <= szy - syz:
-            face_tau = tau
-            face_num = 6
-
-    if face_num > 0:
-        tend = face_tau / axyz
-        tstart = 1.0 - tend
-        return (
-            face_num,
-            [
-                tstart * start[0] + tend * end[0],
-                tstart * start[1] + tend * end[1],
-                tstart * start[2] + tend * end[2]
-            ]
-        )
-    else:
-        return 0, []
-
-################################
 # Matrix/linear algebra math
 ################################
 def pretty(value: float | ArrayLike, *, _depth: int = 0, _shape: Shape | None = None) -> str:
